+++ src/main/java/webserver/server/DispatcherServlet.java
@@ -1,20 +1,22 @@
-package webserver;
+package webserver.server;
 
 import controller.Controller;
 import controller.ForwardController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.http.HttpRequest;
+import webserver.http.HttpResponse;
 
 import java.io.IOException;
 
 public class DispatcherServlet  {
 
-    private final RequestMapping requestMapping = new RequestMapping();
+    private final RequestMapper requestMapper = new RequestMapper();
 
     private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
     protected void service(HttpRequest req, HttpResponse resp) throws IOException {
         logger.info("DispatcherServlet service");
-        Controller controller = requestMapping.getController(req);
+        Controller controller = requestMapper.getController(req);
         String toUrl;
         if(controller == null) {
             controller = new ForwardController(req.getUrl());
